backend/requirements.txt
celery
configparser
Django==1.11.2
django-celery-beat
django-filter
django-cors-headers
djangorestframework
djangorestframework-queryfields
nltk
pymongo
redis
requests[security]
stevedore
altair
SQLAlchemy
django-extensions
scikit-learn
mysqlclient
joblib
scipy
gunicorn
greenlet
gevent